fix ACT and fix fontdata
