+++ LMeter/ACT/ACTClient.cs
@@ -12,18 +12,10 @@ using System.Text.Json;
 using System.Text.Json.Serialization;
 using System.Threading;
 using System.Threading.Tasks;
+using LMeter.ACT.DataStructures;
 
 namespace LMeter.ACT
 {
-    public enum ConnectionStatus
-    {
-        NotConnected,
-        Connected,
-        ShuttingDown,
-        Connecting,
-        ConnectionFailed
-    }
-
     public class ACTClient : IPluginDisposable
     {
         private ACTConfig _config;
@@ -33,6 +25,15 @@ namespace LMeter.ACT
         private ACTEvent? _lastEvent;
         private ConnectionStatus _status;
         private List<ACTEvent> _pastEvents;
+
+        private static readonly JsonSerializerOptions JsonSerializerOptions = new JsonSerializerOptions
+        {
+            PropertyNameCaseInsensitive = true,
+            Converters =
+            {
+                new JsonStringEnumConverter()
+            }
+        };
 
         public static ConnectionStatus Status => Singletons.Get<ACTClient>()._status;
         public static List<ACTEvent> PastEvents => Singletons.Get<ACTClient>()._pastEvents;
@@ -172,13 +173,7 @@ namespace LMeter.ACT
                             {
                                 try
                                 {
-                                    ACTEvent? newEvent = JsonSerializer.Deserialize<ACTEvent>(data, new JsonSerializerOptions
-                                    {
-                                        Converters =
-                                        {
-                                            new JsonStringEnumConverter()
-                                        }
-                                    });
+                                    ACTEvent? newEvent = JsonSerializer.Deserialize<ACTEvent>(data, JsonSerializerOptions);
 
                                     if (newEvent?.Encounter is not null &&
                                         newEvent?.Combatants is not null &&
