--- LMeter/ACT/ACTEvent.cs
@@ -1,289 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
-using LMeter.Enums;
-using System.Text.Json.Serialization;
-
-namespace LMeter.ACT
-{
-    public class ACTEvent
-    {
-        [JsonIgnore]
-        private bool _parsedActive = false;
-
-        [JsonIgnore]
-        private bool _active = false;
-
-        [JsonIgnore]
-        public DateTime Timestamp;
-
-        [JsonPropertyName("type")]
-        public string EventType = string.Empty;
-        
-        [JsonPropertyName("isActive")]
-        public string IsActive = string.Empty;
-        
-        [JsonPropertyName("Encounter")]
-        public Encounter? Encounter;
-        
-        [JsonPropertyName("Combatant")]
-        public Dictionary<string, Combatant>? Combatants;
-
-        public bool IsEncounterActive()
-        {
-            if (_parsedActive)
-            {
-                return _active;
-            }
-            
-            bool.TryParse(this.IsActive, out _active);
-            _parsedActive = true;
-            return _active;
-        }
-
-        public static ACTEvent GetTestData()
-        {
-            return new ACTEvent()
-            {
-                Encounter = LMeter.ACT.Encounter.GetTestData(),
-                Combatants = Combatant.GetTestData()
-            };
-        }
-    }
-
-    public class Encounter
-    {
-        [JsonIgnore]
-        public static string[] TextTags { get; } = typeof(Encounter).GetFields().Select(x => $"[{x.Name.ToLower()}]").ToArray();
-
-        [JsonIgnore]
-        private static readonly Random _rand = new Random();
-        
-        [JsonIgnore]
-        private static readonly Dictionary<string, MemberInfo> _members = typeof(Encounter).GetMembers().ToDictionary((x) => x.Name.ToLower());
-
-        public string GetFormattedString(string format, string numberFormat)
-        {
-            return TextTagFormatter.TextTagRegex.Replace(format, new TextTagFormatter(this, numberFormat, _members).Evaluate);
-        }
-
-        [JsonPropertyName("title")]
-        public string Title = string.Empty;
-
-        [JsonPropertyName("duration")]
-        public string Duration = string.Empty;
-
-        [JsonPropertyName("DURATION")]
-        private string _duration = string.Empty;
-        
-        [JsonPropertyName("encdps")]
-        [JsonConverter(typeof(LazyFloatConverter))]
-        public LazyFloat? Dps;
-
-        [JsonPropertyName("damage")]
-        [JsonConverter(typeof(LazyFloatConverter))]
-        public LazyFloat? DamageTotal;
-        
-        [JsonPropertyName("enchps")]
-        [JsonConverter(typeof(LazyFloatConverter))]
-        public LazyFloat? Hps;
-
-        [JsonPropertyName("healed")]
-        [JsonConverter(typeof(LazyFloatConverter))]
-        public LazyFloat? HealingTotal;
-
-        [JsonPropertyName("damagetaken")]
-        [JsonConverter(typeof(LazyFloatConverter))]
-        public LazyFloat? DamageTaken;
-
-        [JsonPropertyName("deaths")]
-        public string? Deaths;
-
-        [JsonPropertyName("kills")]
-        public string? Kills;
-
-        public static Encounter GetTestData()
-        {
-            float damage = _rand.Next(212345 * 8);
-            float healing = _rand.Next(41234 * 8);
-
-            return new Encounter()
-            {
-                Duration = "00:30",
-                Title = "Preview",
-                Dps = new LazyFloat(damage / 30),
-                Hps = new LazyFloat(healing / 30),
-                Deaths = "0",
-                DamageTotal = new LazyFloat(damage),
-                HealingTotal = new LazyFloat(healing)
-            };
-        }
-    }
-
-    public class Combatant
-    {
-        [JsonIgnore]
-        public static string[] TextTags { get; } = typeof(Combatant).GetFields().Select(x => $"[{x.Name.ToLower()}]").ToArray();
-
-        [JsonIgnore]
-        private static readonly Random _rand = new Random();
-
-        [JsonIgnore]
-        private static readonly Dictionary<string, MemberInfo> _members = typeof(Combatant).GetMembers().ToDictionary((x) => x.Name.ToLower());
-
-        public string GetFormattedString(string format, string numberFormat)
-        {
-            return TextTagFormatter.TextTagRegex.Replace(format, new TextTagFormatter(this, numberFormat, _members).Evaluate);
-        }
-
-        [JsonPropertyName("name")]
-        public string OriginalName = string.Empty;
-
-        [JsonPropertyName("nameOverwrite")] 
-        public string? NameOverwrite = null;
-
-        public string Name => NameOverwrite ?? OriginalName;
-
-        [JsonIgnore]
-        public LazyString<string?>? Name_First;
-
-        [JsonIgnore]
-        public LazyString<string?>? Name_Last;
-
-        [JsonIgnore]
-        public string Rank = string.Empty;
-
-        [JsonPropertyName("job")]
-        public Job Job;
-
-        [JsonIgnore]
-        public LazyString<Job>? JobName;
-
-        [JsonPropertyName("duration")]
-        public string Duration = string.Empty;
-        
-        [JsonPropertyName("encdps")]
-        [JsonConverter(typeof(LazyFloatConverter))]
-        public LazyFloat? EncDps;
-
-        [JsonPropertyName("dps")]
-        [JsonConverter(typeof(LazyFloatConverter))]
-        public LazyFloat? Dps;
-
-        [JsonPropertyName("damage")]
-        [JsonConverter(typeof(LazyFloatConverter))]
-        public LazyFloat? DamageTotal;
-
-        [JsonPropertyName("damage%")]
-        public string DamagePct = string.Empty;
-
-        [JsonPropertyName("crithit%")]
-        public string CritHitPct = string.Empty;
-
-        [JsonPropertyName("DirectHitPct")]
-        public string DirectHitPct = string.Empty;
-
-        [JsonPropertyName("CritDirectHitPct")]
-        public string CritDirectHitPct = string.Empty;
-        
-        [JsonPropertyName("enchps")]
-        [JsonConverter(typeof(LazyFloatConverter))]
-        public LazyFloat? EncHps;
-        
-        [JsonPropertyName("hps")]
-        [JsonConverter(typeof(LazyFloatConverter))]
-        public LazyFloat? Hps;
-
-        public LazyFloat? EffectiveHealing;
-
-        [JsonPropertyName("healed")]
-        [JsonConverter(typeof(LazyFloatConverter))]
-        public LazyFloat? HealingTotal;
-
-        [JsonPropertyName("healed%")]
-        public string HealingPct = string.Empty;
-
-        [JsonPropertyName("overHeal")]
-        [JsonConverter(typeof(LazyFloatConverter))]
-        public LazyFloat? OverHeal;
-
-        [JsonPropertyName("OverHealPct")]
-        public string OverHealPct = string.Empty;
-
-        [JsonPropertyName("damagetaken")]
-        [JsonConverter(typeof(LazyFloatConverter))]
-        public LazyFloat? DamageTaken;
-
-        [JsonPropertyName("deaths")]
-        public string Deaths = string.Empty;
-
-        [JsonPropertyName("kills")]
-        public string Kills = string.Empty;
-
-        [JsonPropertyName("maxhit")]
-        public string MaxHit = string.Empty;
-
-        [JsonPropertyName("MAXHIT")]
-        private string _maxHit = string.Empty;
-
-        public LazyString<string?> MaxHitName;
-
-        public LazyFloat? MaxHitValue;
-
-        public Combatant()
-        {
-            this.Name_First = new LazyString<string?>(() => this.Name, LazyStringConverters.FirstName);
-            this.Name_Last = new LazyString<string?>(() => this.Name, LazyStringConverters.LastName);
-            this.JobName = new LazyString<Job>(() => this.Job, LazyStringConverters.JobName);
-            this.EffectiveHealing = new LazyFloat(() => (this.HealingTotal?.Value ?? 0) - (this.OverHeal?.Value ?? 0));
-            this.MaxHitName = new LazyString<string?>(() => this.MaxHit, LazyStringConverters.MaxHitName);
-            this.MaxHitValue = new LazyFloat(() => LazyStringConverters.MaxHitValue(this.MaxHit));
-        }
-
-        public static Dictionary<string, Combatant> GetTestData()
-        {
-            Dictionary<string, Combatant> mockCombatants = new Dictionary<string, Combatant>();
-            mockCombatants.Add("1", GetCombatant("GNB", "DRK", "WAR", "PLD"));
-            mockCombatants.Add("2", GetCombatant("GNB", "DRK", "WAR", "PLD"));
-
-            mockCombatants.Add("3", GetCombatant("WHM", "AST", "SCH", "SGE"));
-            mockCombatants.Add("4", GetCombatant("WHM", "AST", "SCH", "SGE"));
-
-            mockCombatants.Add("5", GetCombatant("SAM", "DRG", "MNK", "NIN", "RPR"));
-            mockCombatants.Add("6", GetCombatant("SAM", "DRG", "MNK", "NIN", "RPR"));
-            mockCombatants.Add("7", GetCombatant("BLM", "SMN", "RDM"));
-            mockCombatants.Add("8", GetCombatant("DNC", "MCH", "BRD"));
-
-            return mockCombatants;
-        }
-
-        private static Combatant GetCombatant(params string[] jobs)
-        {
-            int damage = _rand.Next(212345);
-            int healing = _rand.Next(41234);
-
-            return new Combatant()
-            {
-                OriginalName = "Firstname Lastname",
-                Duration = "00:30",
-                Job = Enum.Parse<Job>(jobs[_rand.Next(jobs.Length)]),
-                DamageTotal = new LazyFloat(damage.ToString()),
-                Dps = new LazyFloat((damage / 30).ToString()),
-                EncDps = new LazyFloat((damage / 30).ToString()),
-                HealingTotal = new LazyFloat(healing.ToString()),
-                OverHeal = new LazyFloat(5000),
-                Hps = new LazyFloat((healing / 30).ToString()),
-                EncHps = new LazyFloat((healing / 30).ToString()),
-                DamagePct = "100%",
-                HealingPct = "100%",
-                CritHitPct = "20%",
-                DirectHitPct = "25%",
-                CritDirectHitPct = "5%",
-                DamageTaken = new LazyFloat((damage / 20).ToString()),
-                Deaths = _rand.Next(2).ToString(),
-                MaxHit = "Full Thrust-42069"
-            };
-        }
-    }
-}
