+++ LMeter/ACT/ConnectionStatus.cs
@@ -1,0 +1,10 @@
+﻿namespace LMeter.ACT;
+
+public enum ConnectionStatus
+{
+    NotConnected,
+    Connected,
+    ShuttingDown,
+    Connecting,
+    ConnectionFailed
+}
