+++ LMeter/Helpers/FontData.cs
@@ -1,12 +1,15 @@
-﻿namespace LMeter.Helpers;
+﻿using System.Text.Json.Serialization;
+
+namespace LMeter.Helpers;
 
 public struct FontData
 {
-    public string Name;
-    public int Size;
-    public bool Chinese;
-    public bool Korean;
+    public string Name { get; }
+    public int Size { get; }
+    public bool Chinese { get; }
+    public bool Korean { get; }
 
+    [JsonConstructor]
     public FontData(string name, int size, bool chinese, bool korean)
     {
         Name = name;
